refactor(src): reorganize into module directories
Move source files into logical module directories:
- src/common/     — gguf_loader, coreml_code_predictor
- src/tokenizer/  — text_tokenizer
- src/transformer/ — tts_transformer
- src/encoder/    — audio_tokenizer_encoder
- src/decoder/    — audio_tokenizer_decoder
- src/pipeline/   — qwen3_tts, qwen3tts_c_api
- src/main.cpp    — CLI entry point (stays at root)

All includes updated to module-qualified paths (e.g.
"common/gguf_loader.h", "pipeline/qwen3_tts.h").
CMakeLists.txt updated for new file locations.
