+++ src/common/coreml_code_predictor_stub.cpp
@@ -1,4 +1,4 @@
-#include "coreml_code_predictor.h"
+#include "common/coreml_code_predictor.h"
 
 namespace qwen3_tts {
