+++ src/common/gguf_loader.h
