+++ src/pipeline/qwen3tts_c_api.h
