+++ tests/test_transformer.cpp
@@ -1,4 +1,4 @@
-#include "tts_transformer.h"
+#include "transformer/tts_transformer.h"
 
 #include <cstdio>
 #include <cstdlib>
