+++ tests/test_vq_only.cpp
@@ -1,4 +1,4 @@
-#include "audio_tokenizer_decoder.h"
+#include "decoder/audio_tokenizer_decoder.h"
 #include "ggml.h"
 #include "ggml-backend.h"
 #include "ggml-cpu.h"
